+++ CHANGELOG.md
@@ -8,6 +8,7 @@
    - `api_hallo_post`
  - rename APIs
    - `api_get_external_all` to `api_select_data` with url: `/api/[data:class]/[table:method]/[:table_name]`
+   - `api_get_external` to `api_select_my_data` with url: `/api/[my:user_mode]/[data:class]/[table:method]/[:table_name]`
 
 # v1.0.2
 ### New Features